update fixing api datatype
