+++ src/api/User.ts
@@ -11,7 +11,27 @@ export interface UserPostEdit {
   email: string;
 }
 
-export interface getUserDetail extends User {
+export interface getUserDetail extends UserPostEdit {
+  id: string;
+  create_at: Date;
+  update_at: Date;
+  user_type: string;
+  profile_url: string;
+  images: Image[];
+  product_amount: number;
+  business_amount: number;
+  business: Business[];
+  product: Product[];
+}
+
+export interface UserEditResponse {
+  access_token(access_token: any): unknown;
+  phone_number: string;
+  address: string;
+  occupation: string;
+  description: string;
+  username: string;
+  email: string;
   id: string;
   create_at: Date;
   update_at: Date;
