--- src/api/getBusinessAll.ts
@@ -1,14 +1,0 @@
-// import baseApi from "./baseApi";
-
-export interface AllBusiness {
-  business_name: string;
-  business_types: string;
-  create_at: Date;
-  description: string;
-  id: string;
-  profile_url: string;
-  update_at: Date;
-  user_email: string;
-  user_phone_number: string;
-  username: string;
-}
