--- src/api/getProductAll.ts
@@ -1,15 +1,0 @@
-// import baseApi from "./baseApi";
-
-export interface AllProduct {
-  business_id: string;
-  business_name: string;
-  create_at: Date;
-  description: string;
-  id: string;
-  product_name: string;
-  product_price: number;
-  profile_url: string;
-  update_at: Date;
-  user_id: string;
-  username: string;
-}
